新建项目

+++ database/migrations/2019_01_09_141324_create_catalogs_table.php
@@ -19,8 +19,8 @@ class CreateCatalogsTable extends Migration
             $table->integer('item_id')->comment('分项id');
             $table->string('unit')->comment('单位');
             $table->string('mea_mode')->comment('计量方式');
-            $table->decimal('tax_ratio',10,4)->comment('税率');
-            $table->decimal('fees_ratio',10,4)->comment('规费费率');
+            $table->double('tax_ratio')->comment('税率');
+            $table->double('fees_ratio')->comment('规费费率');
             $table->string('remark_one')->comment('备注1');
             $table->string('remark_two')->comment('备注2');
         });
